Update vpc-dns.md
Some customers think that enabling DNS Hostname will generate a public DNS hostname for all instances, even the ones that do not have public IP address, and this will be a security issue. I think we can better clarify this and explain that public DNS hostname will be generated only for instances with EIP associated, and also explain that this option is required for other services like VPC endpoints or private zones to create private hostnames.