Update flow-logs.md
I made some suggestions for the paragraph below.   I think "delay interval" is a better description than "additional time". "Typically" is more appropriate than the word "could".  

If these changes are not accepted, I suggest that an "of" is added to the last sentence in the paragraph - 
In some cases, your logs might be delayed beyond the 5 to 10 minutes "of" additional time mentioned previously.

"After data is captured within an aggregation interval, it takes additional time to process and publish the data to CloudWatch Logs or Amazon S3. This additional time could be around 5 minutes to publish to CloudWatch Logs, and around 10 minutes to publish to Amazon S3. The flow logs service delivers within this additional time in a best effort manner. In some cases, your logs might be delayed beyond the 5 to 10 minutes additional time mentioned previously."