Update flow-logs.md
"After you create a flow log" is better than "After you've created a flow log". Staying in the present tense is always best. 
The phrase "for that network interface" doesn't make sense when you are talking about multiple network interfaces.  The word "that" is problem. 
I used the singular "instance" to simplify the sentence, the reader can easily infer that the activity applies to all new instances. 

Thank you for taking the time to review my proposed changes. I think my first suggestion is mainly a preference while the second suggestion addresses a legitimate grammar issue with the word "that".